Вынес state в компонент App и пробросил в Home & Account
