+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\index.js":"1","D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\App.js":"2","D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\regisration\\registration.js":"3","D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\home\\home.js":"4","D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\account\\account.js":"5","D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\blockTextTitle\\blockTextTitle.js":"6","D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\home_test\\home_test.js":"7","D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\home\\home2.js":"8"},{"size":286,"mtime":1609334670042,"results":"9","hashOfConfig":"10"},{"size":817,"mtime":1610211475400,"results":"11","hashOfConfig":"10"},{"size":1299,"mtime":1609580533692,"results":"12","hashOfConfig":"10"},{"size":9546,"mtime":1610130131114,"results":"13","hashOfConfig":"10"},{"size":4358,"mtime":1610215106540,"results":"14","hashOfConfig":"10"},{"size":424,"mtime":1609573741411,"results":"15","hashOfConfig":"10"},{"size":5130,"mtime":1609787681045,"results":"16","hashOfConfig":"10"},{"size":9027,"mtime":1610217473777,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"115aodf",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"27"},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"34"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\index.js",[],["37","38"],"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\App.js",[],"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\regisration\\registration.js",[],"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\home\\home.js",[],["39","40"],"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\account\\account.js",[],"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\blockTextTitle\\blockTextTitle.js",[],"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\home_test\\home_test.js",[],["41","42"],"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\home\\home2.js",[],{"ruleId":"43","replacedBy":"44"},{"ruleId":"45","replacedBy":"46"},{"ruleId":"43","replacedBy":"47"},{"ruleId":"45","replacedBy":"48"},{"ruleId":"43","replacedBy":"49"},{"ruleId":"45","replacedBy":"50"},"no-native-reassign",["51"],"no-negated-in-lhs",["52"],["51"],["52"],["51"],["52"],"no-global-assign","no-unsafe-negation"]
+[{"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\index.js":"1","D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\App.js":"2","D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\regisration\\registration.js":"3","D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\home\\home.js":"4","D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\account\\account.js":"5","D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\blockTextTitle\\blockTextTitle.js":"6","D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\home_test\\home_test.js":"7","D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\home\\home2.js":"8","D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\blockText\\blockText.js":"9"},{"size":286,"mtime":1609334670042,"results":"10","hashOfConfig":"11"},{"size":3103,"mtime":1610285089449,"results":"12","hashOfConfig":"11"},{"size":1299,"mtime":1609580533692,"results":"13","hashOfConfig":"11"},{"size":9546,"mtime":1610130131114,"results":"14","hashOfConfig":"11"},{"size":4208,"mtime":1610284321484,"results":"15","hashOfConfig":"11"},{"size":424,"mtime":1609573741411,"results":"16","hashOfConfig":"11"},{"size":5130,"mtime":1609787681045,"results":"17","hashOfConfig":"11"},{"size":8752,"mtime":1610284866079,"results":"18","hashOfConfig":"11"},{"size":822,"mtime":1610274137175,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"115aodf",{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"29"},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"36"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\index.js",[],["41","42"],"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\App.js",[],"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\regisration\\registration.js",[],"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\home\\home.js",[],["43","44"],"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\account\\account.js",[],"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\blockTextTitle\\blockTextTitle.js",[],"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\home_test\\home_test.js",[],["45","46"],"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\home\\home2.js",[],"D:\\ALGORITMIKA\\Проекты\\DIPLOM\\pill\\src\\components\\blockText\\blockText.js",[],{"ruleId":"47","replacedBy":"48"},{"ruleId":"49","replacedBy":"50"},{"ruleId":"47","replacedBy":"51"},{"ruleId":"49","replacedBy":"52"},{"ruleId":"47","replacedBy":"53"},{"ruleId":"49","replacedBy":"54"},"no-native-reassign",["55"],"no-negated-in-lhs",["56"],["55"],["56"],["55"],["56"],"no-global-assign","no-unsafe-negation"]
